don't try to destroy master

+++ scripts/destroy.sh
@@ -2,8 +2,13 @@
 
 workspace=$1
 
+if [ "$workspace" == "master" ]; then
+    echo "Will not delete a master deployment"
+    exit 1
+fi
+
 terraform init || exit 1
 terraform workspace select $workspace || terraform workspace new $workspace
-terraform destroy --auto-approve -var="is_teardown=true"
+terraform destroy --auto-approve
 terraform workspace select default
 terraform workspace delete $workspace
